fix inference results

+++ scripts/inference.py
@@ -46,7 +46,7 @@ def inference(method='lstm'):
         for idx, (label, text, offsets) in enumerate(tqdm(inference_dataloader)):
             predicted_label = model(text, offsets)
             predicted_label = (predicted_label > 0.5).squeeze(dim=-1)
-            all_preds += [predicted_label.detach().cpu().numpy()]
+            all_preds += [predicted_label.detach().cpu().numpy().astype(int)]
     all_preds = np.concatenate(all_preds, axis=0)
     df[PARAMS['label']] = all_preds
     return df
